[Refacto] widget 폴더구조 수정
- view들을 각 파일로 나눠 담았습니다
